phase: convert inventory to a phase
