+++ src/game/phase/phase.h
@@ -11,6 +11,7 @@ typedef enum PHASE {
     PHASE_PAUSE,
     PHASE_PICTURE,
     PHASE_STATS,
+    PHASE_INVENTORY,
 } PHASE;
 
 typedef struct PHASER {
@@ -18,7 +19,9 @@ typedef struct PHASER {
     void (*end)();
     int32_t (*control)(int32_t nframes);
     void (*draw)(void);
+    int32_t (*wait)(void);
 } PHASER;
 
+PHASE Phase_Get(void);
 void Phase_Set(PHASE phase, void *arg);
 GAMEFLOW_OPTION Phase_Run();
