+++ src/game/phase/phase_inventory.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "game/phase/phase.h"
+
+extern PHASER g_InventoryPhaser;
